+++ include/models/WSN_flow_model_1.h
@@ -18,7 +18,7 @@ private:
 
     IloNumVar T;
 
-    int M;
+    double M;
 
     virtual void add_objective_function();
 
@@ -54,7 +54,7 @@ WSN_flow_model_1::WSN_flow_model_1(WSN_data &instance) : WSN(instance, "FlowMode
                                                          t(IloNumVarArray(env, instance.n, 0, IloInfinity, ILOFLOAT)),
                                                          l(IloNumVarArray(env, instance.n, 0, IloInfinity, ILOFLOAT)), // Node current formulation
                                                          T(IloNumVar(env, 0, IloInfinity, ILOFLOAT)),
-                                                         M(int(calculates_big_M()))
+                                                         M(calculates_big_M())
 {
 }
