+++ include/models/WSN_mtz_model_2_1.h
@@ -20,7 +20,7 @@ private:
     IloNumVar T;
 
     int p;
-    int M;
+    double M;
 
     void add_mtz_model_variables();
 
@@ -56,7 +56,7 @@ WSN_mtz_model_2_1::WSN_mtz_model_2_1(WSN_data &instance) : WSN(instance, "MTZMod
                                                            pi(IloNumVarArray(env, instance.n, 0, IloInfinity, ILOFLOAT)),
                                                            T(IloNumVar(env, 0, IloInfinity, ILOFLOAT)),
                                                            p((instance.n - instance.number_trees + 1) / 2),
-                                                           M(int(calculates_big_M()))
+                                                           M(calculates_big_M())
 {
 }
